+++ modules/motions/IMURotation.h
@@ -20,7 +20,7 @@ class IMURotation : public Rotation {
    * @param _isClockwise  回頭方向 true:時計回り, false:反時計回り
    * @param _anglePidGain 角度制御PIDゲイン
    */
-  IMURotation(Robot& _robot, int _targetAngle, double _basePower, bool _isClockwise,
+  IMURotation(Robot& _robot, int _targetAngle, int _basePower, bool _isClockwise,
               const PidGain& _anglePidGain);
   /**
    * @brief 回頭する
@@ -53,7 +53,7 @@ class IMURotation : public Rotation {
  private:
   static constexpr float TOLERANCE = 1.0f;  // 許容誤差
   int targetAngle;                          // 目標回転角度(deg) 0~360
-  double basePower;                         // 基準パワー値
+  int basePower;                            // 基準パワー値
   Pid anglePid;                             // 角度PID制御クラス
   float currentAngle;                       // 現在の回頭角度
   double angleError;                        // 角度誤差
